move slider code to useSlider custom hook
